+++ Clean Swift/Router.xctemplate/___FILEBASENAME___Router.swift
@@ -11,14 +11,15 @@
 
 import UIKit
 
-protocol ___FILEBASENAMEASIDENTIFIER___RouterInput {
+protocol ___FILEBASENAMEASIDENTIFIER___RouterInput
+{
   func navigateToSomewhere()
 }
 
 class ___FILEBASENAMEASIDENTIFIER___Router: ___FILEBASENAMEASIDENTIFIER___RouterInput {
   weak var viewController: ___FILEBASENAMEASIDENTIFIER___ViewController!
   
-  // MARK: Navigation
+  // MARK: - Navigation
   
   func navigateToSomewhere() {
     // NOTE: Teach the router how to navigate to another scene. Some examples follow:
@@ -38,7 +39,7 @@ class ___FILEBASENAMEASIDENTIFIER___Router: ___FILEBASENAMEASIDENTIFIER___Router
     // viewController.navigationController?.pushViewController(someWhereViewController, animated: true)
   }
   
-  // MARK: Communication
+  // MARK: - Communication
   
   func passDataToNextScene(segue: UIStoryboardSegue) {
     // NOTE: Teach the router which scenes it can communicate with