+++ Clean Swift/Scene.xctemplate/UICollectionViewController/___FILEBASENAME___Worker.swift
@@ -12,7 +12,7 @@
 import UIKit
 
 class ___FILEBASENAMEASIDENTIFIER___Worker {
-  // MARK: Business Logic
+  // MARK: - Business Logic
   
   func doSomeWork() {
     // NOTE: Do the work