+++ Clean Swift/Scene.xctemplate/UITableViewController/___FILEBASENAME___Interactor.swift
@@ -12,20 +12,20 @@
 import UIKit
 
 protocol ___FILEBASENAMEASIDENTIFIER___InteractorInput {
-  func doSomething(request: ___FILEBASENAMEASIDENTIFIER___Request)
+  func doSomething(request: ___FILEBASENAMEASIDENTIFIER___.Something.Request)
 }
 
 protocol ___FILEBASENAMEASIDENTIFIER___InteractorOutput {
-  func presentSomething(response: ___FILEBASENAMEASIDENTIFIER___Response)
+  func presentSomething(response: ___FILEBASENAMEASIDENTIFIER___.Something.Response)
 }
 
 class ___FILEBASENAMEASIDENTIFIER___Interactor: ___FILEBASENAMEASIDENTIFIER___InteractorInput {
   var output: ___FILEBASENAMEASIDENTIFIER___InteractorOutput!
   var worker: ___FILEBASENAMEASIDENTIFIER___Worker!
   
-  // MARK: Business logic
+  // MARK: - Business logic
   
-  func doSomething(request: ___FILEBASENAMEASIDENTIFIER___Request) {
+  func doSomething(request: ___FILEBASENAMEASIDENTIFIER___.Something.Request) {
     // NOTE: Create some Worker to do the work
     
     worker = ___FILEBASENAMEASIDENTIFIER___Worker()
@@ -33,7 +33,7 @@ class ___FILEBASENAMEASIDENTIFIER___Interactor: ___FILEBASENAMEASIDENTIFIER___In
     
     // NOTE: Pass the result to the Presenter
     
-    let response = ___FILEBASENAMEASIDENTIFIER___Response()
+    let response = ___FILEBASENAMEASIDENTIFIER___.Something.Response()
     output.presentSomething(response: response)
   }
 }