+++ Clean Swift/Scene.xctemplate/UITableViewController/___FILEBASENAME___Presenter.swift
@@ -12,22 +12,22 @@
 import UIKit
 
 protocol ___FILEBASENAMEASIDENTIFIER___PresenterInput {
-  func presentSomething(response: ___FILEBASENAMEASIDENTIFIER___Response)
+  func presentSomething(response: ___FILEBASENAMEASIDENTIFIER___.Something.Response)
 }
 
 protocol ___FILEBASENAMEASIDENTIFIER___PresenterOutput: class {
-  func displaySomething(viewModel: ___FILEBASENAMEASIDENTIFIER___ViewModel)
+  func displaySomething(viewModel: ___FILEBASENAMEASIDENTIFIER___.Something.ViewModel)
 }
 
 class ___FILEBASENAMEASIDENTIFIER___Presenter: ___FILEBASENAMEASIDENTIFIER___PresenterInput {
   weak var output: ___FILEBASENAMEASIDENTIFIER___PresenterOutput!
   
-  // MARK: Presentation logic
+  // MARK: - Presentation logic
   
-  func presentSomething(response: ___FILEBASENAMEASIDENTIFIER___Response) {
+  func presentSomething(response: ___FILEBASENAMEASIDENTIFIER___.Something.Response) {
     // NOTE: Format the response from the Interactor and pass the result back to the View Controller
     
-    let viewModel = ___FILEBASENAMEASIDENTIFIER___ViewModel()
+    let viewModel = ___FILEBASENAMEASIDENTIFIER___.Something.ViewModel()
     output.displaySomething(viewModel: viewModel)
   }
 }