+++ Clean Swift/Unit Tests.xctemplate/___FILEBASENAME___InteractorTests.swift
@@ -13,11 +13,11 @@
 import XCTest
 
 class ___FILEBASENAMEASIDENTIFIER___InteractorTests: XCTestCase {
-  // MARK: Subject under test
+  // MARK: - Subject under test
   
   var sut: ___FILEBASENAMEASIDENTIFIER___Interactor!
   
-  // MARK: Test lifecycle
+  // MARK: - Test lifecycle
   
   override func setUp() {
     super.setUp()
@@ -28,15 +28,15 @@ class ___FILEBASENAMEASIDENTIFIER___InteractorTests: XCTestCase {
     super.tearDown()
   }
   
-  // MARK: Test setup
+  // MARK: - Test setup
   
   func setup___FILEBASENAMEASIDENTIFIER___Interactor() {
     sut = ___FILEBASENAMEASIDENTIFIER___Interactor()
   }
   
-  // MARK: Test doubles
+  // MARK: - Test doubles
   
-  // MARK: Tests
+  // MARK: - Tests
   
   func testSomething() {
     // Given