+++ Clean Swift/Unit Tests.xctemplate/___FILEBASENAME___WorkerTests.swift
@@ -13,11 +13,11 @@
 import XCTest
 
 class ___FILEBASENAMEASIDENTIFIER___WorkerTests: XCTestCase {
-  // MARK: Subject under test
+  // MARK: - Subject under test
   
   var sut: ___FILEBASENAMEASIDENTIFIER___Worker!
   
-  // MARK: Test lifecycle
+  // MARK: - Test lifecycle
   
   override func setUp() {
     super.setUp()
@@ -28,17 +28,18 @@ class ___FILEBASENAMEASIDENTIFIER___WorkerTests: XCTestCase {
     super.tearDown()
   }
   
-  // MARK: Test setup
+  // MARK: - Test setup
   
   func setup___FILEBASENAMEASIDENTIFIER___Worker() {
     sut = ___FILEBASENAMEASIDENTIFIER___Worker()
   }
   
-  // MARK: Test doubles
+  // MARK: - Test doubles
   
-  // MARK: Tests
+  // MARK: - Tests
   
-  func testSomething() {
+  func testSomething()
+  {
     // Given
     
     // When